feat: camera transformation

+++ src/types.h
@@ -38,8 +38,8 @@ struct Camera {
 	double field_of_view; // between 0 and PI
 	double focal_length;
 	double lens_radius; // for depth of field
-	Vector3f view_up = {0,1,0};
-	Vector3f gaze_direction = {0,0,-1};
+	Vector3d view_up = {0,1,0};
+	Vector3d gaze_direction = {0,0,-1};
 };
 
 struct Material {
